cleanup, attempt #2
