--- packages/SwingSet/src/devices/loopbox.js
@@ -1,8 +1,0 @@
-export function buildLoopbox() {
-  const srcPath = require.resolve('./loopbox-src');
-
-  return {
-    srcPath,
-    endowments: {},
-  };
-}
