--- packages/SwingSetRunner/bin/runner
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-require = require('esm')(module);
-const { main } = require('../src/main.js');
-main();
